requirements.txt
Flask-SQLAlchemy==2.4.1
freezegun==0.3.12
pandas==0.25.3
PostgreSQL-Audit==0.10.0
psycopg2==2.7.7
pytest==5.2.2
SQLAlchemy==1.3.11
psycopg2-binary==2.7.6.1
freezegun==0.3.12
requests==2.23.0